Update unit tests to verify set_inp/out_map functionality
